v0.7

+++ fpsHorror/Assets/Game/Scripts/playerScripts/itemEqupe.cs
@@ -1,40 +1,44 @@
 using System.Collections;
 using System.Collections.Generic;
 using TMPro;
+using Unity.Netcode;
 using UnityEngine;
 
-public class itemEqupe : MonoBehaviour
+public class itemEqupe : NetworkBehaviour
 {
     public inventorrySystem system;
     [SerializeField] private TMP_Text BilgiText;
     void Update()
     {
-        BilgiText.text = "";
-        RaycastHit hit;
-        if (Physics.Raycast(transform.position, transform.forward, out hit, 2.5f))
-        {
-            if (hit.collider.gameObject.tag == "Item")
+        if (IsOwner){
+            BilgiText.text = "";
+            RaycastHit hit;
+            if (Physics.Raycast(transform.position, transform.forward, out hit, 2.5f))
             {
-                BilgiText.text = "E equpe";
-                if (Input.GetKeyDown(KeyCode.E))
+                if (hit.collider.gameObject.tag == "Item")
                 {
-                    itemProp prp = hit.collider.gameObject.GetComponent<itemProp>();
-                    system.changeInventory(prp, hit.collider.gameObject);
-                }
-            }
-            else if (hit.collider.gameObject.tag == "sellZone")
-            {
-                
-                if (system.bb00SlotIsFull)
-                {
-                    BilgiText.text = "E to sell";
+                    BilgiText.text = "E equpe";
                     if (Input.GetKeyDown(KeyCode.E))
                     {
-                        system.deleteItem();
+                        itemProp prp = hit.collider.gameObject.GetComponent<itemProp>();
+                        system.changeInventory(prp, hit.collider.gameObject);
                     }
                 }
+                else if (hit.collider.gameObject.tag == "sellZone")
+                {
 
+                    if (system.bb00SlotIsFull)
+                    {
+                        BilgiText.text = "E to sell";
+                        if (Input.GetKeyDown(KeyCode.E))
+                        {
+                            system.deleteItem();
+                        }
+                    }
+
+                }
             }
+        
         }
     }
 }
